Reflection: Fix PHP < 7.1 compatibility
