+++ tests/unit/Codeception/Util/ReflectionTestClass.php
@@ -1,55 +1,8 @@
 <?php
 namespace Codeception\Util;
 
-class ReflectionTestClass
-{
-    const FOO = 'bar';
-
-    private $value = 'test';
-    protected $obj = null;
-    static $flavorOfTheWeek = '';
-
-    public function setInt(int $i): self
-    {
-        $this->value = (string)$i;
-
-        return $this;
-    }
-
-    public function setMixed($m): self
-    {
-        $this->value = (string)$m;
-
-        return $this;
-    }
-
-    public function setValue(string $s): self
-    {
-        $this->value = $s;
-
-        return $this;
-    }
-
-    public function getValue(): string
-    {
-        return $this->value;
-    }
-
-    public function setDebug(Debug $obj, $flavor = self::FOO): self
-    {
-        $this->obj = $obj;
-        self::$flavorOfTheWeek = $flavor;
-
-        return $this;
-    }
-
-    static public function setFlavor(string $flavor = self::FOO)
-    {
-        self::$flavorOfTheWeek = $flavor;
-    }
-
-    private function getSecret(string $s): string
-    {
-        return sprintf("I'm a %s!", $s);
-    }
+if (PHP_VERSION_ID < 70000) {
+    require_once 'ReflectionTestClass56.php';
+} else {
+    require_once 'ReflectionTestClass70.php';
 }
